==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

import sys
import os
script_dir = os.getcwd()
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))


"""
In summary, this line of code modifies the Python path to include the parent 
directory of the directory containing the current script, allowing Python
 to locate and import modules and packages located in that parent directory. 
 It's commonly used in scripts to ensure that project-level modules and packages 
 are accessible 
regardless of the script's current directory.
"""

from src.data import load_raw_data
from pathlib import Path

# Path().parent.resolve().parent / 'data'

from src.data import load_raw_data

rides = load_raw_data(year=2022)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month, input fetures, larget the better
    step_size=24
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

features


tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR

tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

## Plot
