==> notebooks/11_backfill_feature_store.ipynb <==
import os
from dotenv import load_dotenv
import sys

# streamlit run src/frontend.py
# ##################################### IN ORDER TO IMPORT THE PACKAGGE FOR CONDA ENVIORNMENT
# Get the current directory of the script
script_dir = os.getcwd()

# Navigate up one level to the parent directory
project_dir = os.path.abspath(os.path.join(script_dir, '..'))

# Add the parent directory to sys.path
sys.path.append(project_dir)
# #####################################  FOR CONDA ENVIORNMENT 

# script_dir = os.getcwd()
# sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))

from src.paths import PARENT_DIR

import hopsworks

print(hopsworks.__version__)

# load key-value pairs from .env file located in the parent directory
load_dotenv(PARENT_DIR / '.env')

HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']
HOPSWORKS_PROJECT_NAME = 'taxi_demand_calvin'


from datetime import datetime
import pandas as pd
from src.data import load_raw_data

from_year = 2022
to_year = 2023 #datetime.now().year
print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    
    # download data for the whole year
    rides_one_year = load_raw_data(year)
    
    # append rows
    rides = pd.concat([rides, rides_one_year])

print(f'{len(rides)=:,}')

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

# rides

# IcjMMvdfzZBOlCbE.mGsaeXANM7Y0zDxEwNUIRYkBqoEFLntZnC2oxTCRfbAfM1KGrf9C919b3XrFYECI

# HOPSWORKS_API_KEY='IcjMMvdfzZBOlCbE.mGsaeXANM7Y0zDxEwNUIRYkBqoEFLntZnC2oxTCRfbAfM1KGrf9C919b3XrFYECI'

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 3 # for xgboost model training no

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

ts_data

feature_group.insert(ts_data, write_options={"wait_for_job": False})

import hopsworks
