==> notebooks/13_model_training_pipeline.ipynb <==
# push the model to the hopsworks
%load_ext autoreload
%autoreload 2

# Note:
# Lightgbm works, but xgboost not working
# get the feature group,
# create the feature view. 
# we select all the data from feature group and create the feature view based on the selected data

# Different version in hopsworks will have different data

# import lightgbm

import src.config as config

import hopsworks

# connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# connect to the feature store
feature_store = project.get_feature_store()

# connect to the feature group
feature_group = feature_store.get_feature_group(
    name=config.FEATURE_GROUP_NAME,
    version=config.FEATURE_GROUP_VERSION,
)


# import os
# import sys
# script_dir = os.getcwd()
# sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))

# define feature view (get data from multple feature groups and merge it
 # to get the final data set

# FEATURE_VIEW_NAME = 'time_series_hourly_feature_view'
# FEATURE_VIEW_VERSION = 1

# # create feature view (if it doesn't exist yet)
# # This feature view only uses on feature group, so the query is trivial
# try:
#     # create feature view if it doesn't exist yet
#     feature_store.create_feature_view(
#         name=FEATURE_VIEW_NAME,
#         version=FEATURE_VIEW_VERSION,
#         query=feature_group.select_all()
#     )
# except:
#     print('Feature view already existed. Skip creation.')


# # get feature view
# feature_view = feature_store.get_feature_view(
#     name=FEATURE_VIEW_NAME,
#     version=FEATURE_VIEW_VERSION
# )

# create feature view (if it doesn't exist yet)
# This feature view only uses on feature group, so the query is trivial
try:
    # create feature view if it doesn't exist yet
    feature_store.create_feature_view(
        name=config.FEATURE_VIEW_NAME,
        version=config.FEATURE_VIEW_VERSION,
        query=feature_group.select_all()
    )
except:
    print('Feature view already existed. Skip creation.')


# get feature view
feature_view = feature_store.get_feature_view(
    name=config.FEATURE_VIEW_NAME,
    version=config.FEATURE_VIEW_VERSION
)

ts_data, _ = feature_view.training_data(
    description='Time-series hourly taxi rides',
)

# drop `pickup_ts` column
# ts_data.drop('pickup_ts', axis=1, inplace=True)

# sort by `pickup_location_id` and `pickup_hour`
ts_data.sort_values(by=['pickup_location_id', 'pickup_hour'], inplace=True)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28, # one month
    step_size=23,
)

features_and_target = features.copy()
features_and_target['target_rides_next_hour'] = targets

print(f'{features_and_target.shape=}')

from datetime import date, timedelta
from pytz import timezone
import pandas as pd
from src.data_split import train_test_split

# training data -> from January 2022 up until 2 months ago
# test data -> last 2 months
cutoff_date = pd.to_datetime(date.today() - timedelta(days=28*1), utc=True)

print(f'{cutoff_date=}')

X_train, y_train, X_test, y_test = train_test_split(
    features_and_target,
    cutoff_date,
    target_column_name='target_rides_next_hour'   
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

# X_train[X_train.select_dtypes(include=['float32']).columns] = \
#     X_train.select_dtypes(include=['float32']).astype(float)

# y_train = y_train.astype(float)

import numpy as np
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
import optuna

from src.model import get_pipeline

def objective(trial: optuna.trial.Trial) -> float:
    """
    Given a set of hyper-parameters, it trains a model and computes an average
    validation error based on a TimeSeriesSplit
    """
    # pick hyper-parameters
    hyperparams = {
        "metric": 'mae',
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),   
    }
    
    # sort X_train by `pikup_hour` inplace
    # so the TimeSeriesSplit will split the data in a consistent way
    X_train.sort_values('pickup_hour', inplace=True)

    tss = TimeSeriesSplit(n_splits=2)
    scores = []
    for train_index, val_index in tss.split(X_train):

        # split data for training and validation
        X_train_, X_val_ = X_train.iloc[train_index, :], X_train.iloc[val_index,:]
        y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]
        
        # train the model
        pipeline = get_pipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(X_val_)
        mae = mean_absolute_error(y_val_, y_pred)

        scores.append(mae)
   
    # Return the mean score
    return np.array(scores).mean()

study = optuna.create_study(direction="minimize")
study.optimize(objective, n_trials=1)

best_params = study.best_trial.params
print(f'{best_params=}')

# retrain the model based on the best parameter

pipeline = get_pipeline(**best_params)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

import joblib
from src.paths import MODELS_DIR

joblib.dump(pipeline, MODELS_DIR / 'model.pkl')

# X_train[X_train.select_dtypes(include=['float32']).columns] = \
#     X_train.select_dtypes(include=['float32']).astype(float)

# y_train = y_train.astype(float)

X_train.dtypes

from hsml.schema import Schema
from hsml.model_schema import ModelSchema

input_schema = Schema(X_train)
output_schema = Schema(y_train)
model_schema = ModelSchema(input_schema=input_schema, output_schema=output_schema)

# push the model to registrymodel_registry = project.get_model_registry()

model_registry = project.get_model_registry()

model = model_registry.sklearn.create_model(
    name="taxi_demand_predictor_next_hour",
    metrics={"test_mae": test_mae},
    description="Xgboost regressor with a bit of hyper-parameter tuning",
    input_example= X_train.sample(),
    model_schema=model_schema
)



model.save(MODELS_DIR / 'model.pkl')
